Renaming RowMergerV2 to be v2.RowMerger. More v2 classes will be coming.  A v2 package is going to be easier to maintain.
